sms verification: pull out postToVerificationServer
